+++ tests/tests_django_peertube_runner_connector/test_socket.py
@@ -1,7 +1,7 @@
 """Tests for the "socket.py" file of the django_peertube_runner_connector app"""
-from unittest.mock import AsyncMock, patch
+from unittest import mock
 
-from django.test import TestCase
+from django.test import TestCase, override_settings
 
 from django_peertube_runner_connector.factories import RunnerFactory
 from django_peertube_runner_connector.socket import (
@@ -18,8 +18,8 @@ class TestSocket(TestCase):
         """Create a video."""
         self.runner = RunnerFactory()
 
-    @patch("django_peertube_runner_connector.socket.logger")
-    @patch("django_peertube_runner_connector.socket.sio")
+    @mock.patch("django_peertube_runner_connector.socket.logger")
+    @mock.patch("django_peertube_runner_connector.socket.sio")
     async def test_socket_connect(self, _mock_sio, mock_logger):
         """Known runner should be able to connect to the server."""
         await connect(45115, None, {"runnerToken": self.runner.runnerToken})
@@ -28,8 +28,10 @@ class TestSocket(TestCase):
             "Runner with token %s connected with sid %s", self.runner.runnerToken, 45115
         )
 
-    @patch("django_peertube_runner_connector.socket.logger")
-    @patch("django_peertube_runner_connector.socket.sio", new_callable=AsyncMock)
+    @mock.patch("django_peertube_runner_connector.socket.logger")
+    @mock.patch(
+        "django_peertube_runner_connector.socket.sio", new_callable=mock.AsyncMock
+    )
     async def test_socket_connect_wrong_runner_token(self, mock_sio, mock_logger):
         """Known runner should be able to connect to the server."""
         await connect(45115, None, {"runnerToken": "Wrong Runner Token"})
@@ -39,17 +41,68 @@ class TestSocket(TestCase):
         )
         mock_sio.disconnect.assert_called_with(45115, namespace="/runners")
 
-    @patch("django_peertube_runner_connector.socket.logger")
-    @patch("django_peertube_runner_connector.socket.sio", new_callable=AsyncMock)
+    @mock.patch("django_peertube_runner_connector.socket.logger")
+    @mock.patch(
+        "django_peertube_runner_connector.socket.sio", new_callable=mock.AsyncMock
+    )
     async def test_socket_disconnect(self, _mock_sio, mock_logger):
         """Known runner should be able to connect to the server."""
         await disconnect(45115)
 
         mock_logger.info.assert_called_with("%s disconnected", 45115)
 
-    @patch("django_peertube_runner_connector.socket.sio", new_callable=AsyncMock)
-    async def test_socket_send_available_jobs_ping(self, mock_sio):
-        """Known runner should be able to connect to the server."""
+    @mock.patch(
+        "django_peertube_runner_connector.socket.sio", new_callable=mock.AsyncMock
+    )
+    async def test_socket_send_available_jobs_ping_with_no_manager(self, mock_sio):
+        """server with no manager should directly use sio to emit a message."""
         await send_available_jobs_ping_to_runners()
 
         mock_sio.emit.assert_called_with("available-jobs", namespace="/runners")
+
+    @mock.patch(
+        "django_peertube_runner_connector.socket.sio", new_callable=mock.AsyncMock
+    )
+    @override_settings(
+        DJANGO_PEERTUBE_RUNNER_CONNECTOR_SENTINELS=[("localhost", 26379)]
+    )
+    @override_settings(DJANGO_PEERTUBE_RUNNER_CONNECTOR_SENTINELS_MASTER="mymaster")
+    async def test_socket_send_available_jobs_ping_with_sentinel_manager(
+        self, _mock_sio
+    ):
+        """
+        When a sentinel is used as client manager, this one should be used to emit a message
+        """
+        mock_manager = mock.AsyncMock()
+        mock_manager.emit = mock.AsyncMock()
+
+        with mock.patch(
+            "django_peertube_runner_connector.socketio.manager.AsyncSentinelRedisManager",
+            return_value=mock_manager,
+        ):
+            await send_available_jobs_ping_to_runners()
+
+        mock_manager.emit.assert_awaited_once_with(
+            "available-jobs", data=None, namespace="/runners"
+        )
+
+    @mock.patch(
+        "django_peertube_runner_connector.socket.sio", new_callable=mock.AsyncMock
+    )
+    @override_settings(
+        DJANGO_PEERTUBE_RUNNER_CONNECTOR_REDIS="redis://localhost:6379/0"
+    )
+    async def test_socket_send_available_jobs_ping_with_redis_manager(self, _mock_sio):
+        """When redis is used as client manager, this one should be used to emit a message"""
+        mock_manager = mock.AsyncMock()
+        mock_manager.emit = mock.AsyncMock()
+
+        with mock.patch(
+            "django_peertube_runner_connector.socketio.manager.AsyncRedisManager",
+            return_value=mock_manager,
+        ):
+            await send_available_jobs_ping_to_runners()
+
+        mock_manager.emit.assert_awaited_once_with(
+            "available-jobs", data=None, namespace="/runners"
+        )